[Decomp][GameObject] Cohesion split: extract capture-point + destructible to GameObjectDestructible.cpp
